# Use JSON.generate() instead to improve performance