@mtsmfm 2025-05-12
I confirmed it's 1.25x faster on my end, thanks!